# docs/swagger.yaml
basePath: /api/v1
definitions:
  ads-mrkt_internal_analytics_application_http_model.HistoryResponse:
    properties:
      avg_listings_per_user:
        items:
          type: number
        type: array
      commission_earned_ton:
        items:
          type: number
        type: array
      deal_amounts_by_status_ton:
        additionalProperties:
          items:
            format: float64
            type: number
          type: array
        type: object
      deals_by_status:
        additionalProperties:
          items:
            format: int64
            type: integer
          type: array
        type: object
      deals_count:
        items:
          type: integer
        type: array
      from:
        type: string
      listings_count:
        items:
          type: integer
        type: array
      period:
        type: string
      timestamps:
        items:
          type: integer
        type: array
      to:
        type: string
      users_count:
        items:
          type: integer
        type: array
    type: object
  ads-mrkt_internal_analytics_application_http_model.LatestSnapshotResponse:
    properties:
      snapshot:
        $ref: '#/definitions/ads-mrkt_internal_analytics_application_http_model.SnapshotResponse'
    type: object
  ads-mrkt_internal_analytics_application_http_model.SnapshotResponse:
    properties:
      avg_listings_per_user:
        type: number
      commission_earned_ton:
        type: number
      deal_amounts_by_status_ton:
        additionalProperties:
          format: float64
          type: number
        type: object
      deals_by_status:
        additionalProperties:
          format: int64
          type: integer
        type: object
      deals_count:
        type: integer
      id:
        type: integer
      listings_count:
        type: integer
      recorded_at:
        type: string
      users_count:
        type: integer
    type: object
  ads-mrkt_internal_market_application_market_http_model.AuthUserRequest:
    properties:
      referrer:
        type: integer
    type: object
...
  ads-mrkt_internal_market_application_market_http_model.SetWalletRequest:
    properties:
      wallet_address:
        type: string
    type: object
...
  ads-mrkt_internal_market_domain_entity.Channel:
    properties:
      id:
        type: integer
      photo:
        type: string
      title:
        type: string
      username:
        type: string
    type: object
...
  ads-mrkt_internal_server_templates_response.Template:
    properties:
      data: {}
      error_code:
        type: string
      ok:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger for Ads Market API
  version: "1.0"
paths:
  /analytics/snapshot/history:
    get:
      parameters:
      - description: 'Time range: week (7d), month (30d), year (365d)'
        enum:
        - week
        - month
        - year
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Timestamps (Unix) and metric series
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  $ref: '#/definitions/ads-mrkt_internal_analytics_application_http_model.HistoryResponse'
              type: object
        "400":
          description: Bad request (invalid period)
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal error
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      summary: Get analytics snapshot history for charts
      tags:
      - Analytics
  /analytics/snapshot/latest:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Latest snapshot (snapshot field null when none)
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  $ref: '#/definitions/ads-mrkt_internal_analytics_application_http_model.LatestSnapshotResponse'
              type: object
        "500":
          description: Internal error
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      summary: Get latest analytics snapshot
      tags:
      - Analytics
  /market/auth:
    post:
      consumes:
      - application/json
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.AuthUserRequest'
      - description: Telegram init data
        in: header
        name: X-Telegram-InitData
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security: []
      summary: Authenticate user
      tags:
      - Market
  /market/channels/{id}/refresh:
    get:
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Channel
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  $ref: '#/definitions/ads-mrkt_internal_market_domain_entity.Channel'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not found
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Refresh (get) channel data by id; user must be admin of the channel
      tags:
      - Market
  /market/channels/{id}/stats:
    get:
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Channel stats (Telegram broadcast stats shape)
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not found
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Get channel statistics. Allowed for channel admins or users who have
        a listing with this channel.
      tags:
      - Market
...
  /market/me/wallet:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Disconnect wallet (clear linked TON wallet for current user).
      tags:
      - Market
    put:
      consumes:
      - application/json
      parameters:
      - description: wallet_address (raw)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.SetWalletRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Set current user's TON wallet (raw format) for deal payouts.
      tags:
      - Market
  /market/my-channels:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: List of my channels
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ads-mrkt_internal_market_domain_entity.Channel'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: List channels where current user is admin
      tags:
      - Market
...
schemes:
- http
security:
- JWT: []
securityDefinitions:
  JWT:
    description: Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"